+++ packages/eth-block-tracker/tests/withBlockTracker.ts
@@ -1,8 +1,4 @@
-import { providerFromEngine } from '@metamask/eth-json-rpc-provider';
-import type {
-  // Eip1193Request,
-  SafeEventEmitterProvider,
-} from '@metamask/eth-json-rpc-provider';
+import { InternalProvider } from '@metamask/eth-json-rpc-provider';
 import { JsonRpcEngine } from '@metamask/json-rpc-engine';
 import type { Json } from '@metamask/utils';
 import util from 'util';
@@ -16,7 +12,7 @@ type WithPollingBlockTrackerOptions = {
 };
 
 type WithPollingBlockTrackerCallback = (args: {
-  provider: SafeEventEmitterProvider;
+  provider: InternalProvider;
   blockTracker: PollingBlockTracker;
 }) => void | Promise<void>;
 
@@ -98,7 +94,7 @@ function getFakeProvider({
     });
   }
 
-  const provider = providerFromEngine(new JsonRpcEngine());
+  const provider = new InternalProvider({ engine: new JsonRpcEngine() });
   jest
     .spyOn(provider, 'request')
     .mockImplementation(async (eip1193Request): Promise<Json> => {