+++ packages/eth-json-rpc-middleware/src/retryOnEmpty.ts
@@ -1,5 +1,5 @@
 import type { PollingBlockTracker } from '@metamask/eth-block-tracker';
-import type { SafeEventEmitterProvider } from '@metamask/eth-json-rpc-provider';
+import type { InternalProvider } from '@metamask/eth-json-rpc-provider';
 import type { JsonRpcMiddleware } from '@metamask/json-rpc-engine';
 import { createAsyncMiddleware } from '@metamask/json-rpc-engine';
 import type { Json, JsonRpcParams } from '@metamask/utils';
@@ -21,11 +21,7 @@ import { timeout } from './utils/timeout';
 const log = createModuleLogger(projectLogger, 'retry-on-empty');
 // empty values used to determine if a request should be retried
 // `<nil>` comes from https://github.com/ethereum/go-ethereum/issues/16925
-const emptyValues: (string | null | undefined)[] = [
-  undefined,
-  null,
-  '\u003cnil\u003e',
-];
+const emptyValues = [null, '\u003cnil\u003e'];
 
 /**
  * Creates a middleware that retries requests with empty responses.
@@ -39,7 +35,7 @@ export function createRetryOnEmptyMiddleware({
   provider,
   blockTracker,
 }: {
-  provider?: SafeEventEmitterProvider;
+  provider?: InternalProvider;
   blockTracker?: PollingBlockTracker;
 } = {}): JsonRpcMiddleware<JsonRpcParams, Json> {
   if (!provider) {