+++ packages/network-controller/src/create-network-client.ts
@@ -12,9 +12,8 @@ import {
   createFetchMiddleware,
   createRetryOnEmptyMiddleware,
 } from '@metamask/eth-json-rpc-middleware';
-import type { SafeEventEmitterProvider } from '@metamask/eth-json-rpc-provider';
 import {
-  providerFromEngine,
+  InternalProvider,
   providerFromMiddleware,
 } from '@metamask/eth-json-rpc-provider';
 import {
@@ -174,7 +173,7 @@ export function createNetworkClient({
 
   engine.push(networkMiddleware);
 
-  const provider = providerFromEngine(engine);
+  const provider = new InternalProvider({ engine });
 
   const destroy = () => {
     // TODO: Either fix this lint violation or explain why it's necessary to ignore.
@@ -208,7 +207,7 @@ function createBlockTracker({
   getOptions: (
     rpcEndpointUrl: string,
   ) => Omit<PollingBlockTrackerOptions, 'provider'>;
-  provider: SafeEventEmitterProvider;
+  provider: InternalProvider;
 }) {
   const testOptions =
     process.env.IN_TEST && networkClientType === NetworkClientType.Custom
@@ -240,7 +239,7 @@ function createInfuraNetworkMiddleware({
 }: {
   blockTracker: PollingBlockTracker;
   network: InfuraNetworkType;
-  rpcProvider: SafeEventEmitterProvider;
+  rpcProvider: InternalProvider;
   rpcApiMiddleware: JsonRpcMiddleware<JsonRpcParams, Json>;
 }) {
   return mergeMiddleware([
@@ -267,8 +266,6 @@ function createNetworkAndChainIdMiddleware({
   network: InfuraNetworkType;
 }) {
   return createScaffoldMiddleware({
-    // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-    // eslint-disable-next-line @typescript-eslint/naming-convention
     eth_chainId: ChainId[network],
   });
 }