+++ packages/network-controller/tests/network-client/no-block-param.ts
@@ -149,9 +149,7 @@ export function testsForRpcMethodAssumingNoBlockParam(
     });
   });
 
-  for (const emptyValue of [null, undefined, '\u003cnil\u003e']) {
-    // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-    // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+  for (const emptyValue of [null, '\u003cnil\u003e']) {
     it(`does not retry an empty response of "${emptyValue}"`, async () => {
       const request = { method };
       const mockResult = emptyValue;
@@ -175,8 +173,6 @@ export function testsForRpcMethodAssumingNoBlockParam(
       });
     });
 
-    // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-    // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
     it(`does not reuse the result of a previous request if it was "${emptyValue}"`, async () => {
       const requests = [{ method }, { method }];
       const mockResults = [emptyValue, 'some result'];